# primitives/ed25519: Optimize for very large batches